Another try at fixing dllexport/dllimport on Windows.
Patrick Spendrin explains:

_WIN32 is a common define for both mingw and msvc compilers. Exports
can be applied for both compilers on the windows platform.
zip_EXPORTS is a define from cmake on *all* platforms which tells
you that your library is currently built. The standard is
libraryname_EXPORTS, you can redefine it to whatever you want. The
point is that on windows you need use dllexport when the library
gets build and dllimport when it is used (so above code couldn't
have worked on Visual Studio).

Patch from Patrick Spendrin.